Diego/charting (#772)
Used the highcharts library to copy the chart in https://jmerle.github.io/battlecode-2024-statistics/ . This is not a 1:1 chart because they used a theme that did everything for them, while I opted to copy it using explicit settings. I put it in the home screen of /home.

The reason the data looks smooth is because its fake data I generated that has many thousand data points. If used on a lot less data points, it looks blocks like the one on the website. I still need to implement fetching data from the API